Refine bench phase event parsing
